some more thoughts about where to put the data:
we can keep it locally in our own folders (and on GDrive) without syncing it to the cloud using this .gitignore file. This means my code for this data will reference this subfolder, but the data itself will stay local. You could add a line to the .gitignore for any subfolder you'd like too (in Bash/Linux: $ echo directory >> .gitignore)

If we're both editing and want to stop messing with pathnames IN the code, we can use the same folder and file titles for the data as a workaround to not having the data on the shared server repo.